Move cql_model traits to own crate
Faucilitates the further breakup of cql_storage
