+++ cql_storage/Cargo.toml
@@ -5,5 +5,6 @@ authors = ["Andrew Sisley"]
 edition = "2018"
 
 [dependencies]
+cql_model = { path = "../cql_model" }
 byteorder = "1"
 time = "0.1"
