+++ cql_storage/src/lib.rs
@@ -2,7 +2,6 @@ pub mod f64;
 pub mod u64;
 pub mod f64_nullable;
 pub mod tiny_text;
-pub mod cql_type;
 
 mod internal {
     pub mod grow_database;
